+++ app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".MyApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -19,22 +20,22 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false">
         </activity>
-        <activity android:name=".login.LoginActivity" android:exported="true">
+        <activity android:name=".presentation.login.LoginActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".signup.SignUpActivity"
+            android:name=".presentation.signup.SignUpActivity"
             android:exported="false">
         </activity>
 
         <activity
-            android:name=".Home.HomeActivity"
+            android:name=".presentation.home.HomeActivity"
             android:exported="true">
         </activity>
