+++ app/src/main/java/org/sopt/dosopttemplate/data/model/request/RequestLoginDto.kt
@@ -1,16 +1,12 @@
-package org.sopt.dosopttemplate.data.signup
+package org.sopt.dosopttemplate.data.model.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignupDto(
+data class RequestLoginDto(
     @SerialName("username")
     val username: String,
     @SerialName("password")
     val password: String,
-    @SerialName("major")
-    val major: String,
-    @SerialName("nickname")
-    val nickname: String,
 )
