+++ app/src/main/java/org/sopt/dosopttemplate/data/model/request/RequestSignupDto.kt
@@ -1,12 +1,16 @@
-package org.sopt.dosopttemplate.data.login
+package org.sopt.dosopttemplate.data.model.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLoginDto(
+data class RequestSignupDto(
     @SerialName("username")
     val username: String,
     @SerialName("password")
     val password: String,
+    @SerialName("major")
+    val major: String,
+    @SerialName("nickname")
+    val nickname: String,
 )
