+++ app/src/main/java/org/sopt/dosopttemplate/data/model/response/ResponseLoginDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.login
+package org.sopt.dosopttemplate.data.model.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
