--- app/src/main/java/org/sopt/dosopttemplate/login/LoginState.kt
@@ -1,8 +1,0 @@
-package org.sopt.dosopttemplate.login
-
-
-sealed class LoginState<out T> {
-    object Loading : LoginState<Nothing>()
-    data class Success<T>(val data: T) : LoginState<T>()
-    object Error : LoginState<Nothing>()
-}
