+++ app/src/main/java/org/sopt/dosopttemplate/presentation/login/LoginViewModel.kt
@@ -1,24 +1,19 @@
-package org.sopt.dosopttemplate.login
+package org.sopt.dosopttemplate.presentation.login
 
 import android.util.Log
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import org.sopt.dosopttemplate.data.api.ServicePool
-import org.sopt.dosopttemplate.data.api.ServicePool.authService
-import org.sopt.dosopttemplate.data.login.RequestLoginDto
-import org.sopt.dosopttemplate.data.login.ResponseLoginDto
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import org.sopt.dosopttemplate.data.model.request.RequestLoginDto
+import org.sopt.dosopttemplate.data.model.response.ResponseLoginDto
+import org.sopt.dosopttemplate.di.ServicePool.authService
 
 class LoginViewModel : ViewModel() {
-    private val _loginState = MutableStateFlow<LoginState<ResponseLoginDto>>(LoginState.Loading)
-    val loginState: StateFlow<LoginState<ResponseLoginDto>> = _loginState.asStateFlow()
+    private val _uiState = MutableStateFlow<UiState<ResponseLoginDto>>(UiState.Loading)
+    val uiState: StateFlow<UiState<ResponseLoginDto>> = _uiState.asStateFlow()
 
     fun login(id: String, password: String) {
         viewModelScope.launch {
@@ -28,16 +23,16 @@ class LoginViewModel : ViewModel() {
                 if (response.isSuccessful) {
                     val responseBody = response.body()
                     if (responseBody != null) {
-                        _loginState.value = LoginState.Success(responseBody)
+                        _uiState.value = UiState.Success(responseBody)
                     } else {
-                        _loginState.value = LoginState.Error
+                        _uiState.value = UiState.Error
                     }
                 } else {
-                    _loginState.value = LoginState.Error
+                    _uiState.value = UiState.Error
                 }
             } catch (e: Exception) {
                 Log.d("LoginViewModel", "실패")
-                _loginState.value = LoginState.Error
+                _uiState.value = UiState.Error
             }
         }
     }
