+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
