+++ app/src/main/java/org/sopt/dosopttemplate/presentation/user/UserAdapter.kt
@@ -1,13 +1,13 @@
-package org.sopt.dosopttemplate.user
+package org.sopt.dosopttemplate.presentation.user
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.dosopttemplate.data.user.ResponseUserData
 import org.sopt.dosopttemplate.databinding.ItemUserBinding
+import org.sopt.dosopttemplate.domain.entity.UserEntity
 
 class UserAdapter : RecyclerView.Adapter<UserViewHolder>() {
-    private val userList = mutableListOf<ResponseUserData>()
+    private val userList = mutableListOf<UserEntity>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UserViewHolder {
         val binding = ItemUserBinding.inflate(LayoutInflater.from(parent.context), parent, false)
@@ -22,7 +22,7 @@ class UserAdapter : RecyclerView.Adapter<UserViewHolder>() {
         return userList.size
     }
 
-    fun setUserList(userData: List<ResponseUserData>) {
+    fun setUserList(userData: List<UserEntity>) {
         userList.clear()
         userList.addAll(userData)
         notifyDataSetChanged()
