+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home.HomeActivity">
+    tools:context=".presentation.home.HomeActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fcv_home"
