+++ app/src/main/res/layout/activity_signup.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="SignupViewModel"
-            type="org.sopt.dosopttemplate.signup.SignupViewModel" />
+            type="org.sopt.dosopttemplate.presentation.signup.SignupViewModel" />
     </data>
 
 <androidx.constraintlayout.widget.ConstraintLayout
@@ -14,7 +14,7 @@
     android:layout_height="match_parent"
     android:paddingHorizontal="60dp"
     android:paddingVertical="30dp"
-    tools:context=".signup.SignUpActivity">
+    tools:context=".presentation.signup.SignUpActivity">
 
     <TextView
         android:id="@+id/welcome"
