+++ app/src/main/res/layout/fragment_do_android.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DoAndroid.DoAndroidFragment">
+    tools:context=".presentation.doandroid.DoAndroidFragment">
 
     <TextView
         android:layout_width="match_parent"
